rename optimizer_utils to onnx_utils
