+++ onnxruntime/test/python/transformers/test_onnx_utils.py
@@ -8,7 +8,7 @@ import numpy
 from onnx import ModelProto, TensorProto, helper
 from onnx.external_data_helper import set_external_data
 
-from onnxruntime.transformers.optimizer_utils import extract_raw_data_from_model, has_external_data
+from onnxruntime.transformers.onnx_utils import extract_raw_data_from_model, has_external_data
 
 
 class TestOptimizerUtils(unittest.TestCase):
